Fix fix fix job

+++ srcs/builtins/built_fg.c
@@ -2,13 +2,14 @@
 #include "shell.h"
 #include "execute.h"
 #include <signal.h>
+#include <errno.h>
 
 void			killpids(void *data)
 {
 	pid_t pid;
 
 	pid = *(pid_t *)data;
-	kill(SIGINT, pid);
+	kill(pid, SIGINT);
 	waitpid(pid, 0, WUNTRACED);
 }
 
@@ -23,23 +24,16 @@ int				built_fg(char **av, char **env)
 
 	if (!vector_get_len(g_jobs))
 		return (0);
-	/* pid = *(pid_t *)vector_back(g_jobs); */
-	/* vector_pop_back(&g_jobs); */
-	/* pgid = getpgid(pid); */
-	/* tcsetpgrp(0, pgid); */
-	/* killpg(pgid, SIGCONT); */
-	/* xwaitpid(pgid, WUNTRACED); */
-	/* tcsetpgrp(0, getpgid(getpid())); */
 	job = vector_back(g_jobs);
 	vector_pop_back(&g_jobs);
 	g_pids = job->pids;
 	pid = *(pid_t *)vector_back(g_pids);
 	pgid = getpgid(pid);
 	tcsetpgrp(0, pgid);
-	/* if (job->state == JOB_STOP) */
+	if (job->state == JOB_STOP)
 		killpg(pgid, SIGCONT);
 	xwaitpid(pid, WUNTRACED);
-	if (!g_pids)
+	if (g_job)
 		return (SHERR_OK);
 	vector_foreach(g_pids, killpids);
 	return (SHERR_OK);
